refactor(l5): move table-projections from src/ into server/domain (issue 864)
Moved (old -> new):

- src/table-projections.ts -> server/domain/table-projections.ts

The file's import of src/types.ts is rewritten relative (../../src/types.ts)
per M3; types.ts itself is not pulled along and belongs to a later lane.

The old path becomes an L5 shim per M2, so both of its importers
(src/rest-api.ts, src/tools/get-entry.ts) keep their import lines untouched
and stay out of the pre-commit gate. The shim retires with src/ at L6.

src/ runtime closure under server/main.ts: 67 -> 66.

Five of the six files this lane was briefed to move stayed in src/, because
each fails the server/ standard M4 requires a moved file to meet whole:

- src/read-policy.ts: its test drives shared-namespace.ts through the
  SHARED_NAMESPACE_CANONICAL, SHARED_NAMESPACE_PHYSICAL, and
  SHARED_NAMESPACE_LEGACY environment keys. server/**/*.ts forbids
  process.env, and the module owning those keys has not moved yet, so M4's
  stated response is to leave the file in src/ and name the blocking keys.
- src/source-refs.ts, src/candidate-dedupe.ts, src/distill-window.ts,
  src/distiller.ts: each carries a test file whose describe() block exceeds
  the 100-line function rule, and whose bodies hold 88 non-null assertions.
  Clearing them means restructuring the suites, which M5 forbids (test
  bodies, names, and assertions stay identical) and which would reach into
  scripts/assert-db-tests-ran.ts. Left for a lane briefed to do that work.

src/extraction.ts was left in src/ as the brief directs: its callers include
scripts/bulk-import.ts and six legacy src/ files, so an options-object
signature change would drag every one of them into the gate.
